change packages naming and folders
