+++ app/process/handler/crypto-transfer/crypto-transfer_test.go
@@ -1,4 +1,4 @@
-package crypto_transfer
+package cryptotransfer
 
 import (
 	"errors"
